Refactor reports repository for MySQL compatibility and improve query performance
- Updated SQL queries to use MySQL syntax (e.g., TIMESTAMPDIFF, DATE_FORMAT).
- Replaced PostgreSQL-specific functions and type casts with MySQL equivalents.
- Enhanced parameter binding in queries for better security and performance.
- Cleaned up code by removing unnecessary type casts and simplifying logic.
- Added debug logging in LeadDetails component for better error tracking.
- Introduced FollowupsDebug component for testing followups API directly.
- Improved error handling in leadsService for fetching followups.d

+++ backend/crm/modules/bookings/bookings.repository.js
@@ -670,8 +670,8 @@ function createBookingsRepository({ db, logger, schema }) {
               SUM(CASE WHEN status = 'PENDING' AND ${notDeletedPredicate} THEN 1 ELSE 0 END) AS pending_bookings,
               SUM(CASE WHEN status = 'COMPLETED' AND ${notDeletedPredicate} THEN 1 ELSE 0 END) AS completed_bookings,
               SUM(CASE WHEN status = 'CANCELLED' AND ${notDeletedPredicate} THEN 1 ELSE 0 END) AS cancelled_bookings,
-              COALESCE(SUM(CASE WHEN status <> 'CANCELLED' AND ${notDeletedPredicate} THEN COALESCE(total_amount, 0) ELSE 0 END), 0)::numeric AS total_revenue,
-              COALESCE(SUM(CASE WHEN status <> 'CANCELLED' AND ${notDeletedPredicate} THEN GREATEST(COALESCE(total_amount, 0) - COALESCE(advance_received, 0), 0) ELSE 0 END), 0)::numeric AS pending_payments_amount,
+              COALESCE(SUM(CASE WHEN status <> 'CANCELLED' AND ${notDeletedPredicate} THEN COALESCE(total_amount, 0) ELSE 0 END), 0) AS total_revenue,
+              COALESCE(SUM(CASE WHEN status <> 'CANCELLED' AND ${notDeletedPredicate} THEN GREATEST(COALESCE(total_amount, 0) - COALESCE(advance_received, 0), 0) ELSE 0 END), 0) AS pending_payments_amount,
               SUM(CASE WHEN status <> 'CANCELLED' AND ${notDeletedPredicate} AND COALESCE(advance_received, 0) < COALESCE(total_amount, 0) THEN 1 ELSE 0 END) AS pending_payments_count
             FROM ${schema.tableName}
           `;
